feat(usecase): validate route id before delete

+++ src/main/kotlin/domain/usecase/crud/route/DeleteRouteUseCase.kt
@@ -1,11 +1,18 @@
 package domain.usecase.crud.route
 
 import domain.repository.RouteRepository
+import domain.validator.RouteValidator
+import domain.validator.ValidationResult
 
 class DeleteRouteUseCase(
     private val routeRepository: RouteRepository
 ) {
     suspend operator fun invoke(id: String): Boolean {
+        val validationResult = RouteValidator.validateId(id)
+        if (validationResult is ValidationResult.Failure) {
+            return false
+        }
+
         return routeRepository.deleteRoute(id)
     }
 }
